+++ spec/helpers/bundler_wrapper_spec.rb
@@ -14,7 +14,8 @@ end
 
 describe "Bundler version sorting" do
   it "sorts the keys correctly" do
-    expect(LanguagePack::Helpers::BundlerWrapper::SORTED_KEYS).to eq(["2.3", "2.4", "2.5", "2.6", "2.7"])
+    expect(LanguagePack::Helpers::BundlerWrapper::SORTED_KEYS)
+      .to eq(["2.3", "2.4", "2.5", "2.6", "2.7", "4.0"])
   end
 end
 
@@ -47,6 +48,15 @@ describe "Bundler version detection" do
 
     version = wrapper_klass.detect_bundler_version(contents: "BUNDLED WITH\n   2.999.7")
     expect(version).to eq(wrapper_klass::BLESSED_BUNDLER_VERSIONS[wrapper_klass::BUNDLER_2_LARGEST])
+
+    version = wrapper_klass.detect_bundler_version(contents: "BUNDLED WITH\n   4.0.0")
+    expect(version).to eq(wrapper_klass::BLESSED_BUNDLER_VERSIONS[wrapper_klass::BUNDLER_4_SMALLEST])
+
+    version = wrapper_klass.detect_bundler_version(contents: "BUNDLED WITH\n   4.0.0.beta2")
+    expect(version).to eq(wrapper_klass::BLESSED_BUNDLER_VERSIONS[wrapper_klass::BUNDLER_4_SMALLEST])
+
+    version = wrapper_klass.detect_bundler_version(contents: "BUNDLED WITH\n   4.999999.0")
+    expect(version).to eq(wrapper_klass::BLESSED_BUNDLER_VERSIONS[wrapper_klass::BUNDLER_4_LARGEST])
 
     expect {
       wrapper_klass.detect_bundler_version(contents: "BUNDLED WITH\n   3.6.7")